requirements.txt
streamlit
groq
python-dotenv
tiktoken
chromadb
sentence-transformers
torch